UPDATE - did a code clean up and moved github extraction to a separate script
